DROID-2378 move constants
