+++ payments/src/main/java/com/anytypeio/anytype/payments/mapping/MembershipExt.kt
@@ -11,13 +11,13 @@ import com.anytypeio.anytype.core_models.membership.MembershipPaymentMethod.METH
 import com.anytypeio.anytype.core_models.membership.MembershipPeriodType
 import com.anytypeio.anytype.core_models.membership.MembershipTierData
 import com.anytypeio.anytype.core_ui.R
-import com.anytypeio.anytype.payments.constants.MembershipConstants
-import com.anytypeio.anytype.payments.constants.MembershipConstants.ACTIVE_TIERS_WITH_BANNERS
-import com.anytypeio.anytype.payments.constants.MembershipConstants.CO_CREATOR_ID
-import com.anytypeio.anytype.payments.constants.MembershipConstants.MEMBERSHIP_CONTACT_EMAIL
-import com.anytypeio.anytype.payments.constants.MembershipConstants.MEMBERSHIP_LEVEL_DETAILS
-import com.anytypeio.anytype.payments.constants.MembershipConstants.PRIVACY_POLICY
-import com.anytypeio.anytype.payments.constants.MembershipConstants.TERMS_OF_SERVICE
+import com.anytypeio.anytype.core_models.membership.MembershipConstants
+import com.anytypeio.anytype.core_models.membership.MembershipConstants.ACTIVE_TIERS_WITH_BANNERS
+import com.anytypeio.anytype.core_models.membership.MembershipConstants.CO_CREATOR_ID
+import com.anytypeio.anytype.core_models.membership.MembershipConstants.MEMBERSHIP_CONTACT_EMAIL
+import com.anytypeio.anytype.core_models.membership.MembershipConstants.MEMBERSHIP_LEVEL_DETAILS
+import com.anytypeio.anytype.core_models.membership.MembershipConstants.PRIVACY_POLICY
+import com.anytypeio.anytype.core_models.membership.MembershipConstants.TERMS_OF_SERVICE
 import com.anytypeio.anytype.payments.models.BillingPriceInfo
 import com.anytypeio.anytype.payments.models.MembershipPurchase
 import com.anytypeio.anytype.payments.models.Tier
